@@ -37,16 +37,16 @@
     "chartist": "^0.11.4",
     "chartist-plugin-tooltips-updated": "^0.1.4",
     "moment-timezone": "^0.5.31",
-    "react": "^16.13.1",
+    "react": "^18.3.1",
     "react-chartist": "^0.14.4",
     "react-copy-to-clipboard": "^5.0.3",
     "react-datetime": "^3.0.4",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.3.1",
     "react-github-btn": "^1.2.0",
     "react-live": "^2.2.3",
     "react-router-dom": "^5.2.0",
     "react-router-hash-link": "^2.3.1",
-    "react-scripts": "3.4.3",
+    "react-scripts": "5.0.1",
     "react-transition-group": "^4.4.1",
     "sass": "^1.50.0",
     "simplebar-react": "^2.3.0"
